requirements.txt
# Core Framework
fastapi==0.115.4
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# Multi-Agent
langgraph==0.2.50
langchain==0.3.9
langchain-openai==0.2.11
langchain-community==0.3.9

# LLM
openai==1.55.3
groq==0.13.0
# RAG / Vector DB
faiss-cpu==1.13.2
sentence-transformers==3.3.1

# Resume Parsing
PyMuPDF==1.24.14
pdfplumber==0.11.4

# Database
pymongo==4.9.1
motor==3.6.0

# Frontend
streamlit==1.40.2
streamlit-extras==0.4.7

# Utilities
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
httpx==0.28.0
aiofiles==24.1.0
numpy==1.26.4
tiktoken==0.8.0

# Logging & Monitoring
loguru==0.7.3

# Optional Speech
# openai-whisper==20231117
# gTTS==2.5.3

-e .